requirements.txt
psycopg2-binary
bcrypt
python-dotenv
flask
flask-cors